// src/app/interfaces/character-info.ts
export interface CharacterInfo {
    cName: { key: String; value: String};
    cLevel: { key: String; value: Number};
    cClass: { key: String; value: String};
    pName: { key: String; value: String};
    cBackground: { key: String; value: String};
    pRace: { key: String; value: String};
    cAlignement: { key: String; value: String};
}